+++ examples/Ladybug/LadybugDSMX/LadybugDSMX.ino
@@ -34,11 +34,11 @@ Copyright (c) 2018 Simon D. Levy
 #include "boards/arduino/ladybug.hpp"
 #include "receivers/arduino/dsmx.hpp"
 #include "mixers/quadxcf.hpp"
-#include "pidcontrollers/acro.hpp"
 #include "pidcontrollers/level.hpp"
 #include "pidcontrollers/yaw.hpp"
+#include "pidcontrollers/acro.hpp"
 
-constexpr uint8_t CHANNEL_MAP[6] = {0, 1, 2, 3, 6, 4};
+const uint8_t CHANNEL_MAP[6] = {0, 1, 2, 3, 6, 4};
 
 hf::Hackflight h;
 
@@ -46,7 +46,7 @@ hf::DSMX_Receiver rc = hf::DSMX_Receiver(CHANNEL_MAP);
 
 hf::MixerQuadXCF mixer;
 
-hf::AcroPid acroPid = hf::AcroPid(0.225, 0.001875, 0.375); 
+hf::AcroPid acroPid = hf::AcroPid(0.225, 0.001875, 0.375);
 
 hf::LevelPid levelPid = hf::LevelPid(0.20);
 
@@ -57,9 +57,10 @@ void setup(void)
     // Initialize Hackflight firmware
     h.init(new hf::Ladybug(), &rc, &mixer);
 
-    // Add Level, Yaw PIDs for aux switch position 0
-    h.addPidController(&yawPid, 0);
-    h.addPidController(&levelPid, 0);
+    // Add Level, Yaw PIDs
+    h.addPidController(&levelPid);
+    h.addPidController(&yawPid);
+    h.addPidController(&acroPid);
 }
 
 void loop(void)
